:truck: [templates] Move prompts adapter from application.cookiecutter
